pyproject.toml
[project]
name = "mimetic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Elliot Hill", email = "elliot.douglas.hill@gmail.com" }]
requires-python = ">=3.13"
dependencies = ["tensordict>=0.11.0", "torch>=2.10.0"]

[dependency-groups]
dev = [
    "marimo>=0.20.4",
    "matplotlib>=3.10.0",
    "myst-parser>=4.0.0",
    "numpydoc>=1.10.0",
    "pydata-sphinx-theme>=0.16.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "ruff>=0.15.5",
    "sphinx>=9.1.0",
    "sphinx-autobuild>=2025.8.25",
    "sphinx-autodoc-typehints>=3.9.8",
    "sphinx-copybutton>=0.5.0",
    "sphinx-design>=0.6.0",
]

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"

[tool.ruff.format]
skip-magic-trailing-comma = true

[tool.numpydoc_validation]
checks = [
    "all",  # enable all checks...
    "EX01", # ...except: requires Examples section (often too strict)
    "SA01", # ...except: requires See Also section
    "ES01", # ...except: requires extended summary
]
exclude = [
    '\.__repr__$',
    '\.__str__$',
    '^test_',      # skip test functions
]
